+++ modules/4337/src/utils/safe.ts
@@ -1,8 +1,11 @@
-import { AddressLike, JsonRpcProvider, Provider, Signer, ethers } from 'ethers'
+import { MultiProvider4337, RpcProvider } from '@safe-global/safe-4337-provider'
+import { Provider, Signer, ethers } from 'ethers'
 
 // Import from Safe contracts repo once it is upgraded to ethers v6 and can be installed via npm
 import { MetaTransaction, SafeSignature, SignedSafeTransaction, buildSignatureBytes } from './execution'
 import { PackedUserOperation, UserOperation, EIP712_SAFE_OPERATION_TYPE, packGasParameters, unpackUserOperation } from './userOp'
+
+export { MultiProvider4337 }
 
 const AddressOne = '0x0000000000000000000000000000000000000001'
 
@@ -89,61 +92,6 @@ const callInterface = async (provider: Provider, contract: string, method: strin
 
 const actionCalldata = (action: MetaTransaction): string => {
   return INTERFACES.encodeFunctionData('executeUserOp', [action.to, action.value, action.data, action.operation])
-}
-
-export interface RpcProvider extends Provider {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  send(method: string, params: unknown[]): Promise<any>
-}
-
-export class MultiProvider4337 extends JsonRpcProvider {
-  generalProvider: RpcProvider
-  constructor(aaProviderUrl: string, generalProvider: RpcProvider) {
-    super(aaProviderUrl)
-    this.generalProvider = generalProvider
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  send(method: string, params: unknown[]): Promise<any> {
-    if (
-      [
-        'eth_supportedEntryPoints',
-        'eth_estimateUserOperationGas',
-        'eth_sendUserOperation',
-        'eth_getUserOperationByHash',
-        'eth_getUserOperationReceipt',
-      ].indexOf(method) >= 0
-    ) {
-      return super.send(method, params)
-    } else {
-      return this.generalProvider.send(method, params)
-    }
-  }
-
-  public async sendUserOperation(userOp: UserOperation, entryPoint: AddressLike): Promise<string> {
-    const jsonUserOp = {
-      sender: ethers.getAddress(userOp.sender),
-      nonce: ethers.toBeHex(userOp.nonce),
-      callData: ethers.hexlify(userOp.callData),
-      callGasLimit: ethers.toBeHex(userOp.callGasLimit),
-      verificationGasLimit: ethers.toBeHex(userOp.verificationGasLimit),
-      preVerificationGas: ethers.toBeHex(userOp.preVerificationGas),
-      maxFeePerGas: ethers.toBeHex(userOp.maxFeePerGas),
-      maxPriorityFeePerGas: ethers.toBeHex(userOp.maxPriorityFeePerGas),
-      signature: ethers.hexlify(userOp.signature),
-    } as Record<string, unknown>
-    if (userOp.factory) {
-      jsonUserOp.factory = ethers.getAddress(userOp.factory)
-      jsonUserOp.factoryData = ethers.hexlify(userOp.factoryData!)
-    }
-    if (userOp.paymaster) {
-      jsonUserOp.paymaster = ethers.getAddress(userOp.paymaster)
-      jsonUserOp.paymasterVerificationGasLimit = ethers.toBeHex(userOp.paymasterVerificationGasLimit!)
-      jsonUserOp.paymasterPostOpGasLimit = ethers.toBeHex(userOp.paymasterPostOpGasLimit!)
-      jsonUserOp.paymasterData = ethers.hexlify(userOp.paymasterData!)
-    }
-    return await super.send('eth_sendUserOperation', [jsonUserOp, await ethers.resolveAddress(entryPoint, this)])
-  }
 }
 
 export class Safe4337Operation {
