+++ modules/4337/test/e2e/UniqueSigner.spec.ts
@@ -1,6 +1,6 @@
+import { bundlerRpc, prepareAccounts, waitForUserOp } from '@safe-global/safe-4337-local-bundler'
 import { expect } from 'chai'
 import { deployments, ethers, network } from 'hardhat'
-import { bundlerRpc, prepareAccounts, waitForUserOp } from '../utils/e2e'
 import { chainId } from '../utils/encoding'
 import { packGasParameters, unpackUserOperation } from '../../src/utils/userOp'
