+++ modules/4337/test/e2e/WebAuthnSingletonSigner.spec.ts
@@ -1,6 +1,7 @@
+import { bundlerRpc, prepareAccounts, waitForUserOp } from '@safe-global/safe-4337-local-bundler'
 import { expect } from 'chai'
 import { deployments, ethers, network } from 'hardhat'
-import { bundlerRpc, encodeMultiSendTransactions, prepareAccounts, waitForUserOp } from '../utils/e2e'
+import { encodeMultiSendTransactions } from '../utils/encoding'
 import {
   UserVerificationRequirement,
   WebAuthnCredentials,
