+++ modules/allowances/package.json
@@ -45,7 +45,6 @@
     "hardhat": "^2.21.0",
     "hardhat-deploy": "^0.12.1",
     "hardhat-gas-reporter": "^1.0.10",
-    "rimraf": "^5.0.5",
     "solhint": "4.1.1",
     "solidity-coverage": "^0.8.7",
     "ts-node": "^10.9.2",
