--- modules/passkey/test/utils/encoding.ts
@@ -1,5 +1,0 @@
-import { ethers } from 'hardhat'
-
-export async function chainId(): Promise<bigint> {
-  return (await ethers.provider.getNetwork()).chainId
-}
