+++ package.json
@@ -4,8 +4,8 @@
   "description": "A monorepo for Safe modules",
   "main": "dist/index.js",
   "workspaces": [
-    "modules/*",
     "packages/*",
+    "modules/*",
     "examples/*"
   ],
   "scripts": {
@@ -39,6 +39,7 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.5",
     "prettier": "^3.2.5",
-    "prettier-plugin-solidity": "^1.3.1"
+    "prettier-plugin-solidity": "^1.3.1",
+    "rimraf": "^5.0.5"
   }
 }
