+++ packages/4337-local-bundler/src/index.ts
@@ -1,6 +1,6 @@
-import 'hardhat-deploy'
-
 import deployEntryPoint from './deploy/entrypoint'
 import deploySafe from './deploy/safe'
+
+export * from './testing'
 
 export { deployEntryPoint, deploySafe }
