+++ packages/4337-local-bundler/src/types/hardhat.ts
@@ -1,0 +1,2 @@
+import '@nomicfoundation/hardhat-ethers'
+import 'hardhat-deploy'
