+++ packages/4337-provider/.eslintignore
@@ -1,0 +1,1 @@
+dist/
