+++ packages/4337-provider/README.md
@@ -1,0 +1,3 @@
+# ERC-4337 Ethers Provider
+
+This package contains an Ethers.js RPC provider with support for interacting with ERC-4337 bundlers.
